Use sigsetjmp/siglongjmp in linux
